Bug: ProductInfo.displayProductInfo is not a function
Let's try to fix this issue
