+++ manifest.json
@@ -33,4 +33,4 @@
     "permissions": [
         "activeTab"
      ]
-  }
+}
